+++ src/main/java/org/board/exercise_board/comment/controller/CommentController.java
@@ -3,11 +3,11 @@ package org.board.exercise_board.comment.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.board.exercise_board.application.DeleteApplication;
 import org.board.exercise_board.application.ModifyApplication;
 import org.board.exercise_board.application.WriteApplication;
 import org.board.exercise_board.comment.domain.dto.CommentDto;
 import org.board.exercise_board.comment.domain.form.CommentForm;
-import org.board.exercise_board.comment.service.CommentService;
 import org.board.exercise_board.user.domain.model.User;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -24,9 +24,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/comment")
 public class CommentController {
 
-  private final CommentService commentService;
   private final WriteApplication writeApplication;
   private final ModifyApplication modifyApplication;
+  private final DeleteApplication deleteApplication;
 
   @Operation(summary = "댓글 작성")
   @PostMapping("/{postId}/write")
@@ -61,7 +61,7 @@ public class CommentController {
       @PathVariable(value = "commentId") Long commentId,
       @AuthenticationPrincipal User user
   ) {
-    return ResponseEntity.ok(commentService.deleteComment(
+    return ResponseEntity.ok(deleteApplication.deleteComment(
         postId, commentId, user.getUsername()));
   }
 }
